split out models and daos
