+++ mini_gplus/daos/exceptions.py
@@ -1,0 +1,13 @@
+from werkzeug.exceptions import HTTPException
+
+
+class UnauthorizedAccess(HTTPException):
+    pass
+
+
+class BadRequest(HTTPException):
+    pass
+
+
+class NotFound(HTTPException):
+    pass
