+++ mini_gplus/daos/make_uuid.py
@@ -1,0 +1,5 @@
+import uuid
+
+
+def make_uuid():
+    return str(uuid.uuid4())
