+++ mini_gplus/models/created_at_mixin.py
@@ -1,0 +1,4 @@
+class CreatedAtMixin(object):
+    @property
+    def created_at(self):
+        return self.id.generation_time.timestamp()
