+++ mini_gplus/models/media.py
@@ -1,0 +1,6 @@
+from mongoengine import Document, StringField
+from .created_at_mixin import CreatedAtMixin
+
+
+class Media(Document, CreatedAtMixin):
+    object_name = StringField(required=True, unique=True)
