+++ mini_gplus/models/reaction.py
@@ -1,0 +1,12 @@
+from mongoengine import Document, ReferenceField, StringField, CASCADE
+from .created_at_mixin import CreatedAtMixin
+from .user import User
+
+
+class Reaction(Document, CreatedAtMixin):
+    eid = StringField(required=True)
+    author = ReferenceField(User, required=True, reverse_delete_rule=CASCADE)  # type: User
+    emoji = StringField(required=True)
+
+    def make_href(self, parent_post):
+        return f"/post/{parent_post.eid}#reaction-{self.eid}"
