+++ mini_gplus/tests/daos/base_test_case.py
@@ -1,0 +1,11 @@
+from unittest import TestCase
+from mongoengine import connect, disconnect
+
+
+class BaseTestCase(TestCase):
+    def setUp(self):
+        self.connection = connect('mongoenginetest', host='mongomock://localhost')
+
+    def tearDown(self):
+        self.connection.drop_database("mongoenginetest")
+        disconnect()
